Fixed segfault on bad exception path
