+++ native/common/jp_bytetype.cpp
@@ -1,5 +1,5 @@
 /*****************************************************************************
-   Copyright 2004-2008 Steve MÃ©nard
+   Copyright 2004-2008 Steve Ménard
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -249,8 +249,8 @@ void JPByteType::releaseView(JPArrayView& view)
 {
 	try
 	{
-	JPJavaFrame frame(view.getContext());
-	frame.ReleaseByteArrayElements((jbyteArray) view.m_Array->getJava(),
+		JPJavaFrame frame(view.getContext());
+		frame.ReleaseByteArrayElements((jbyteArray) view.m_Array->getJava(),
 				(jbyte*) view.m_Memory, view.m_Buffer.readonly ? JNI_ABORT : 0);
 	}	catch (JPypeException& ex)
 	{
