+++ native/common/jp_chartype.cpp
@@ -209,9 +209,16 @@ void JPCharType::getView(JPArrayView& view)
 
 void JPCharType::releaseView(JPArrayView& view)
 {
-	JPJavaFrame frame(view.getContext());
-	frame.ReleaseCharArrayElements((jcharArray) view.m_Array->getJava(),
+	try
+	{
+		JPJavaFrame frame(view.getContext());
+		frame.ReleaseCharArrayElements((jcharArray) view.m_Array->getJava(),
 			(jchar*) view.m_Memory, view.m_Buffer.readonly ? JNI_ABORT : 0);
+	}	catch (JPypeException& ex)
+	{
+		// This is called as part of the cleanup routine and exceptions
+		// are not permitted
+	}
 }
 
 const char* JPCharType::getBufferFormat()
