+++ native/common/jp_doubletype.cpp
@@ -326,9 +326,16 @@ void JPDoubleType::getView(JPArrayView& view)
 
 void JPDoubleType::releaseView(JPArrayView& view)
 {
-	JPJavaFrame frame(view.getContext());
-	frame.ReleaseDoubleArrayElements((jdoubleArray) view.m_Array->getJava(),
+	try
+	{
+		JPJavaFrame frame(view.getContext());
+		frame.ReleaseDoubleArrayElements((jdoubleArray) view.m_Array->getJava(),
 			(jdouble*) view.m_Memory, view.m_Buffer.readonly ? JNI_ABORT : 0);
+	}	catch (JPypeException& ex)
+	{
+		// This is called as part of the cleanup routine and exceptions
+		// are not permitted
+	}
 }
 
 const char* JPDoubleType::getBufferFormat()
