+++ native/common/jp_floattype.cpp
@@ -303,9 +303,16 @@ void JPFloatType::getView(JPArrayView& view)
 
 void JPFloatType::releaseView(JPArrayView& view)
 {
-	JPJavaFrame frame(view.getContext());
-	frame.ReleaseFloatArrayElements((jfloatArray) view.m_Array->getJava(),
+	try
+	{
+		JPJavaFrame frame(view.getContext());
+		frame.ReleaseFloatArrayElements((jfloatArray) view.m_Array->getJava(),
 			(jfloat*) view.m_Memory, view.m_Buffer.readonly ? JNI_ABORT : 0);
+	}	catch (JPypeException& ex)
+	{
+		// This is called as part of the cleanup routine and exceptions
+		// are not permitted
+	}
 }
 
 const char* JPFloatType::getBufferFormat()
