+++ native/common/jp_inttype.cpp
@@ -280,9 +280,16 @@ void JPIntType::getView(JPArrayView& view)
 
 void JPIntType::releaseView(JPArrayView& view)
 {
-	JPJavaFrame frame(view.getContext());
-	frame.ReleaseIntArrayElements((jintArray) view.m_Array->getJava(),
+	try
+	{
+		JPJavaFrame frame(view.getContext());
+		frame.ReleaseIntArrayElements((jintArray) view.m_Array->getJava(),
 			(jint*) view.m_Memory, view.m_Buffer.readonly ? JNI_ABORT : 0);
+	}	catch (JPypeException& ex)
+	{
+		// This is called as part of the cleanup routine and exceptions
+		// are not permitted
+	}
 }
 
 const char* JPIntType::getBufferFormat()
