+++ native/common/jp_longtype.cpp
@@ -280,9 +280,16 @@ void JPLongType::getView(JPArrayView& view)
 
 void JPLongType::releaseView(JPArrayView& view)
 {
-	JPJavaFrame frame(view.getContext());
-	frame.ReleaseLongArrayElements((jlongArray) view.m_Array->getJava(),
+	try
+	{
+		JPJavaFrame frame(view.getContext());
+		frame.ReleaseLongArrayElements((jlongArray) view.m_Array->getJava(),
 			(jlong*) view.m_Memory, view.m_Buffer.readonly ? JNI_ABORT : 0);
+	}	catch (JPypeException& ex)
+	{
+		// This is called as part of the cleanup routine and exceptions
+		// are not permitted
+	}
 }
 
 const char* JPLongType::getBufferFormat()
