+++ test/jpypetest/test_jbyte.py
@@ -1,0 +1,103 @@
+import jpype
+from jpype.types import *
+import sys
+import logging
+import time
+import common
+
+class JByteTestCase(common.JPypeTestCase):
+    def setUp(self):
+        common.JPypeTestCase.setUp(self)
+        self.fixture = jpype.JClass("jpype.common.Fixture")()
+
+    def testByteFromInt(self):
+        self.assertEqual(self.fixture.callByte(int(123)), 123)
+
+    @common.requireNumpy
+    def testByteFromNPInt(self):
+        import numpy as np
+        self.assertEqual(self.fixture.callByte(np.int(123)), 123)
+
+    @common.requireNumpy
+    def testByteFromNPInt8(self):
+        import numpy as np
+        self.assertEqual(self.fixture.callByte(np.int8(123)), 123)
+        self.assertEqual(self.fixture.callByte(np.uint8(123)), 123)
+
+    @common.requireNumpy
+    def testByteFromNPInt16(self):
+        import numpy as np
+        self.assertEqual(self.fixture.callByte(np.int16(123)), 123)
+        self.assertEqual(self.fixture.callByte(np.uint16(123)), 123)
+
+    @common.requireNumpy
+    def testByteFromNPInt32(self):
+        import numpy as np
+        self.assertEqual(self.fixture.callByte(np.int32(123)), 123)
+        self.assertEqual(self.fixture.callByte(np.uint32(123)), 123)
+
+    @common.requireNumpy
+    def testByteFromNPInt64(self):
+        import numpy as np
+        self.assertEqual(self.fixture.callByte(np.int64(123)), 123)
+        self.assertEqual(self.fixture.callByte(np.uint64(123)), 123)
+
+    def testByteFromFloat(self):
+        with self.assertRaises(TypeError):
+            self.fixture.callByte(float(2))
+
+    @common.requireNumpy
+    def testByteFromNPFloat(self):
+        import numpy as np
+        with self.assertRaises(TypeError):
+            self.fixture.callByte(np.float(2))
+
+    @common.requireNumpy
+    def testByteFromNPFloat32(self):
+        import numpy as np
+        with self.assertRaises(TypeError):
+            self.fixture.callByte(np.float32(2))
+
+    @common.requireNumpy
+    def testByteFromNPFloat64(self):
+        import numpy as np
+        with self.assertRaises(TypeError):
+            self.fixture.callByte(np.float64(2))
+
+    def testByteRange(self):
+        with self.assertRaises(OverflowError):
+            self.fixture.callByte(int(1e10))
+        with self.assertRaises(OverflowError):
+            self.fixture.callByte(int(-1e10))
+
+    def testByteFromNone(self):
+        with self.assertRaises(TypeError):
+            self.fixture.callByte(None)
+
+    def testByteArrayAsString(self):
+        t = JClass("jpype.array.TestArray")()
+        v = t.getByteArray()
+        self.assertEqual(str(v), 'avcd')
+
+    def testByteArrayIntoVector(self):
+        ba = jpype.JArray(jpype.JByte)(b'123')
+        v = jpype.java.util.Vector(1)
+        v.add(ba)
+        self.assertEqual(len(v), 1)
+        self.assertNotEqual(v[0], None)
+
+    def testByteArraySimple(self):
+        a = JArray(JByte)(2)
+        a[1] = 2
+        self.assertEqual(a[1], 2)
+
+    def testJArrayConversionByte(self):
+        expected = (0, 1, 2, 3)
+        ByteBuffer = jpype.java.nio.ByteBuffer
+        bb = ByteBuffer.allocate(4)
+        buf = bb.array()
+        for i in range(len(expected)):
+            buf[i] = expected[i]
+        for i in range(len(expected)):
+            self.assertEqual(expected[i], buf[i])
+
